Split typescript spec into page-objects and support
