+++ ee_tests/src/specs/page_objects/base.page.ts
@@ -1,0 +1,14 @@
+import { browser } from 'protractor';
+
+export abstract class BasePage {
+  protected url: string = ''; // Will be same as baseUrl by default.
+
+  constructor(url?: string) {
+    this.url = url || '';
+  }
+
+  async open() {
+    await browser.get(this.url);
+  }
+
+}
