+++ ee_tests/src/specs/page_objects/home.page.ts
@@ -1,0 +1,10 @@
+import { $ } from 'protractor';
+import { BasePage } from './base.page';
+
+export class HomePage extends BasePage {
+  login = $('#login');
+
+  constructor(url: string) {
+    super(url);
+  }
+}
