+++ ee_tests/src/specs/support/index.ts
@@ -29,3 +29,4 @@ export function desktopTestSetup() {
   setBrowserMode(BrowserMode.Desktop);
 }
 
+
